@@ -1,23 +1,20 @@
-import logo from './logo.svg';
-import './App.css';
+import React, { useState } from "react";
+import "./App.css";
+import ListOfGifs from "./components/ListOfGifs";
+
+import { Link, Route } from "wouter";
 
 function App() {
+  //const [keyword, setKeyword] = useState("hi");
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <section className="App-content">
+        <h1>App</h1>
+        <Link to="/gif/panda">Gifs de Pandas</Link>
+        <Link to="/gif/matrix">Gifs de Matrix</Link>
+        <Link to="/gif/transformers">Gifs de Transformers</Link>
+        <Route component={ListOfGifs} path="/gif/:keyword" />
+      </section>
     </div>
   );
 }
